prfix コンポーネント移動
