+++ app/src/App.tsx
@@ -4,7 +4,9 @@ import {
   QueryClient,
   QueryClientProvider,
 } from "@tanstack/react-query";
-import { fetchTasks, postTask } from "./util";
+import { fetchTasks, postTask } from "./api/task";
+import TaskButton from "./components/TaskButton";
+import Task from "./components/Task";
 
 const queryClient = new QueryClient();
 
@@ -32,34 +34,12 @@ const MainPage: React.FC = () => {
         {tasks.data?.tasks.map((task) => (
           <li key={task.id}>
             <Task task={task} />
-            <TaskButton onClick={createTaskMutation.mutate}>
+            <TaskButton onClick={createTaskMutation.mutate} type="button">
               タスクを作成する
             </TaskButton>
           </li>
         ))}
       </ul>
     </div>
-  );
-};
-
-const Task: React.FC<{ task: TaskData }> = ({ task }) => {
-  return (
-    <>
-      <h2>{task.title}</h2>
-      <p>{task.createdAt}</p>
-      <p>{task.finishedAt}</p>
-    </>
-  );
-};
-
-const TaskButton: React.FC<TaskManagementButtonProps> = (props) => {
-  return (
-    <button
-      onClick={() => {
-        props.onClick();
-      }}
-    >
-      {props.children}
-    </button>
   );
 };
