+++ app/src/components/Task.tsx
@@ -1,0 +1,10 @@
+const Task: React.FC<{ task: TaskData }> = ({ task }) => {
+  return (
+    <>
+      <h2>{task.title}</h2>
+      <p>{task.createdAt}</p>
+      <p>{task.finishedAt}</p>
+    </>
+  );
+};
+export default Task;
