+++ app/src/components/TaskButton.tsx
@@ -1,0 +1,20 @@
+type TaskManagementButtonProps = {
+  type: "button" | "submit"; // ボタンの種類
+  onClick: () => void; // 作成、更新、削除の処理を行う関数
+  children: React.ReactNode;
+};
+
+const TaskButton: React.FC<TaskManagementButtonProps> = (props) => {
+  return (
+    <button
+      type={props.type}
+      onClick={() => {
+        props.onClick();
+      }}
+    >
+      {props.children}
+    </button>
+  );
+};
+
+export default TaskButton;
